+++ dessia_common/utils/serialization.py
@@ -22,6 +22,8 @@ from dessia_common.typings import InstanceOf
 from dessia_common.graph import explore_tree_from_leaves  # , cut_tree_final_branches
 from dessia_common.breakdown import get_in_object_from_path
 
+fullargsspec_cache = {}
+
 
 def serialize_dict(dict_):
     """
@@ -200,14 +202,10 @@ def deserialize(serialized_element, sequence_annotation: str = 'List',
 def deserialize_sequence(sequence, annotation=None,
                          global_dict=None, pointers_memo=None,
                          path='#'):
-    # TODO: rename to deserialize sequence? Or is this a duplicate ?
     origin, args = dcty.unfold_deep_annotation(typing_=annotation)
-    deserialized_sequence = []
-    for ie, elt in enumerate(sequence):
-        path_elt = f'{path}/{ie}'
-        deserialized_element = deserialize(elt, args, global_dict=global_dict,
-                                           pointers_memo=pointers_memo, path=path_elt)
-        deserialized_sequence.append(deserialized_element)
+    deserialized_sequence = [deserialize(elt, args, global_dict=global_dict, pointers_memo=pointers_memo,
+                                         path=f'{path}/{ie}') for ie, elt in enumerate(sequence)]
+
     if origin is tuple:
         # Keeping as a tuple
         return tuple(deserialized_sequence)
@@ -250,7 +248,12 @@ def dict_to_object(dict_, class_=None, force_generic: bool = False,
 
             return obj
 
-        class_argspec = inspect.getfullargspec(class_)
+        if class_ in fullargsspec_cache:
+            class_argspec = fullargsspec_cache[class_]
+        else:
+            class_argspec = inspect.getfullargspec(class_)
+            fullargsspec_cache[class_] = class_argspec
+
         init_dict = {k: v for k, v in dict_.items() if k in class_argspec.args + class_argspec.kwonlyargs}
         # TOCHECK Class method to generate init_dict ??
     else: